+++ R/utils-example.R
@@ -51,9 +51,7 @@ head_data <- function(x, n = 1L){
 #' @param x `data.frame`
 #'
 #' @return `list`
-#'
-#' @keywords internal
-#' @export
+#' @noRd
 #'
 flip <- function(x) {
 
@@ -71,9 +69,7 @@ flip <- function(x) {
 #' @param ggspec, ggspec (does this have an S3 class?)
 #'
 #' @return modified copy of `ggspec`
-#'
-#' @keywords internal
-#' @export
+#' @noRd
 #'
 truncate_data_ggspec <- function(ggspec) {
 
@@ -97,9 +93,7 @@ truncate_data_ggspec <- function(ggspec) {
 #' @param vegaspec, vegaspec (does this have an S3 class?)
 #'
 #' @return modified copy of `vegaspec`
-#'
-#' @keywords internal
-#' @export
+#' @noRd
 #'
 truncate_data_vegaspec <- function(vegaspec) {