+++ data-raw/examples/ggplot/barchart-flipped-weighted-normalized-mpg.R
@@ -1,3 +1,0 @@
-  ggplot(data = mpg, aes(x = class, weight = displ)) +
-  geom_bar(aes(fill = drv), position = "fill") +
-  coord_flip()