+++ data-raw/examples/ggplot/barchart-weighted-mpg.R
@@ -1,2 +1,0 @@
-  ggplot(data = mpg, aes(x = class, weight = displ)) +
-  geom_bar()