+++ inst/examples/ggplot/barchart-mpg.R
@@ -1,2 +1,0 @@
-  ggplot(data = mpg, aes(x = class)) +
-  geom_bar()