+++ tests/testthat/examples/ggplot/bar_flip_weight_norm_mpg.R
@@ -1,0 +1,3 @@
+ggplot(head(mpg, 1), aes(x = class, weight = displ)) +
+  geom_bar(aes(fill = drv), position = "fill") +
+  coord_flip()