+++ tests/testthat/examples/ggplot/barchart-weighted-mpg.R
@@ -1,2 +1,0 @@
-  ggplot(head(mpg, 1), aes(x = class, weight = displ)) +
-  geom_bar()