+++ tests/testthat/test-utils-examples.R
@@ -51,9 +51,9 @@ test_that("truncate_data functions work", {
   # test only if the data-raw directory is available
   skip_if_not(fs::dir_exists(here::here("data-raw")))
 
-  example <- "scatterplot-iris"
+  example <- "point_iris"
 
-  gg_scatter <- source(dev_example_path(example, "ggspec"))$value
+  gg_scatter <- source(ggv_dev_path(example, "ggspec"))$value
 
   gg_scatter_truncate <- gg_scatter
   gg_scatter_truncate$data$`data-00`$observations <-
@@ -61,7 +61,7 @@ test_that("truncate_data functions work", {
 
   expect_identical(truncate_data_ggspec(gg_scatter), gg_scatter_truncate)
 
-  vl_scatter <- source(dev_example_path(example, "vegaspec"))$value
+  vl_scatter <- source(ggv_dev_path(example, "vegaspec"))$value
 
   vl_scatter_truncate <- vl_scatter
   vl_scatter_truncate$datasets$`data-00` <-
@@ -80,15 +80,15 @@ test_that(".example_ functions work", {
   skip_if_not(fs::dir_exists(here::here("data-raw")))
 
   names <- .example_names("ggplot", "dev")
-  expect_true("scatterplot-iris" %in% names)
+  expect_true("point_iris" %in% names)
 
   expect_is(
-    .example_obj("scatterplot-iris", type = "ggplot", source = "dev"),
+    .example_obj("point_iris", type = "ggplot", source = "dev"),
     "gg"
   )
 
   expect_is(
-    .example_obj("scatterplot-iris", type = "vegaspec", source = "dev"),
+    .example_obj("point_iris", type = "vegaspec", source = "dev"),
     "vegaspec"
   )