+++ NEWS.md
@@ -1,8 +1,14 @@
 # ggvega  0.0.0 (development version) 
 
-* Added support for `geom_bar()` and `stat_count()` (#42, #44).
+* Introduce `ggvega_examples`: a list of example-objects. Each object has a `plot`, `code`, and `as_vegaspec()`
 
-* Added an option to `as_vegaspec()`, `single_view` to indicate to collapse a single-layer ggplot2 object into a single-view Vega-Lite specification. This makes it easier to add selections and conditional encodings using {vlbuildr} (#70). 
+* Update to Vega-Lite 4.0.2. (#79)
+
+* Internal change to use "modern" V8 engine and Vega-Lite classes directly. (#77)
+
+* Added support for `geom_bar()` and `stat_count()`. (#42, #44)
+
+* Added an option to `as_vegaspec()`, `single_view` to indicate to collapse a single-layer ggplot2 object into a single-view Vega-Lite specification. This makes it easier to add selections and conditional encodings using {vlbuildr}. (#70)
 
 * Added support for `ggplot2::coord_flip()`.